# services/platform-api/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyticsAccepted": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Accepted",
            "type": "integer"
          },
          "retentionDays": {
            "const": 30,
            "default": 30,
            "title": "Retentiondays",
            "type": "integer"
          }
        },
        "required": [
          "accepted"
        ],
        "title": "AnalyticsAccepted",
        "type": "object"
      },
      "AnalyticsAggregate": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "minimum": 5.0,
            "title": "Count",
            "type": "integer"
          },
          "name": {
            "enum": [
              "app_opened",
              "onboarding_completed",
              "scan_started",
              "scan_completed",
              "capture_retake_requested",
              "analysis_completed",
              "analysis_unavailable",
              "comparison_viewed",
              "observation_map_viewed",
              "report_generated",
              "report_downloaded",
              "summary_video_generated",
              "share_created",
              "share_revoked",
              "clinician_review_requested",
              "learning_section_viewed",
              "notification_permission_changed"
            ],
            "title": "Name",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "started",
              "completed",
              "abstained",
              "cancelled",
              "failed",
              "viewed",
              "generated",
              "shared",
              "revoked",
              "enabled",
              "disabled"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "platform": {
            "enum": [
              "ios",
              "android",
              "web"
            ],
            "title": "Platform",
            "type": "string"
          }
        },
        "required": [
          "name",
          "platform",
          "outcome",
          "count"
        ],
        "title": "AnalyticsAggregate",
        "type": "object"
      },
      "AnalyticsBatch": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/AnalyticsEventInput"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "AnalyticsBatch",
        "type": "object"
      },
      "AnalyticsConsentResponse": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "policyVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policyversion"
          },
          "updatedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          }
        },
        "required": [
          "enabled",
          "policyVersion",
          "updatedAt"
        ],
        "title": "AnalyticsConsentResponse",
        "type": "object"
      },
      "AnalyticsConsentUpdate": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "policyVersion": {
            "const": "analytics-v1",
            "default": "analytics-v1",
            "title": "Policyversion",
            "type": "string"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "AnalyticsConsentUpdate",
        "type": "object"
      },
      "AnalyticsEventInput": {
        "additionalProperties": false,
        "properties": {
          "appVersion": {
            "pattern": "^[A-Za-z0-9._+-]{1,32}$",
            "title": "Appversion",
            "type": "string"
          },
          "name": {
            "enum": [
              "app_opened",
              "onboarding_completed",
              "scan_started",
              "scan_completed",
              "capture_retake_requested",
              "analysis_completed",
              "analysis_unavailable",
              "comparison_viewed",
              "observation_map_viewed",
              "report_generated",
              "report_downloaded",
              "summary_video_generated",
              "share_created",
              "share_revoked",
              "clinician_review_requested",
              "learning_section_viewed",
              "notification_permission_changed"
            ],
            "title": "Name",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "started",
              "completed",
              "abstained",
              "cancelled",
              "failed",
              "viewed",
              "generated",
              "shared",
              "revoked",
              "enabled",
              "disabled"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "platform": {
            "enum": [
              "ios",
              "android",
              "web"
            ],
            "title": "Platform",
            "type": "string"
          },
          "surface": {
            "enum": [
              "app",
              "onboarding",
              "scan",
              "result",
              "comparison",
              "map",
              "report",
              "sharing",
              "clinician",
              "learn",
              "settings"
            ],
            "title": "Surface",
            "type": "string"
          }
        },
        "required": [
          "name",
          "platform",
          "appVersion",
          "surface",
          "outcome"
        ],
        "title": "AnalyticsEventInput",
        "type": "object"
      },
      "AnalyticsSummary": {
        "additionalProperties": false,
        "properties": {
          "days": {
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Days",
            "type": "integer"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/AnalyticsAggregate"
            },
            "title": "Groups",
            "type": "array"
          },
          "minimumGroupSize": {
            "const": 5,
            "default": 5,
            "title": "Minimumgroupsize",
            "type": "integer"
          }
        },
        "required": [
          "days",
          "groups",
          "generatedAt"
        ],
        "title": "AnalyticsSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Stoma3D Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/admin/analytics/summary": {
      "get": {
        "operationId": "aggregate_analytics_v2_admin_analytics_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 30,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Aggregate Analytics",
        "tags": [
          "privacy-safe analytics"
        ]
      }
    },
...
    "/v2/analytics/events": {
      "post": {
        "operationId": "submit_analytics_v2_analytics_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Analytics",
        "tags": [
          "privacy-safe analytics"
        ]
      }
    },
...
    "/v2/me/analytics-consent": {
      "get": {
        "operationId": "get_analytics_consent_v2_me_analytics_consent_get",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsConsentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analytics Consent",
        "tags": [
          "privacy-safe analytics"
        ]
      },
      "put": {
        "operationId": "update_analytics_consent_v2_me_analytics_consent_put",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsConsentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsConsentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Analytics Consent",
        "tags": [
          "privacy-safe analytics"
        ]
      }
    },
...
  }
}